requirements.txt
beautifulsoup4
django
djangorestframework
django-haystack
gunicorn
dj-database-url
python-social-auth
requests
elasticsearch
